refactor: api cluster/* and cluster/acme/* root methods were condensed into their own files
